Refactored to home.js
