+++ src/components/index.js
@@ -1,3 +1,3 @@
 export { AccountSummaryComponent } from './AccountSummary'
-export { HomeComponent } from './Home'
+export { HomeComponent } from './Home/home'
 export { SubscriptionComponent } from './Subscription'
